@@ -5,11 +5,11 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
-  cloud_firestore: ^4.0.5
-  firebase_auth: ^4.1.2
+  cloud_firestore: ^4.13.2
+  firebase_auth: ^4.14.1
   flutter:
     sdk: flutter
   flutter_localizations:
