@@ -1,19 +1,19 @@
 name: flutter_support_chat
 description: Flutter package to implement a firestore chat between customer and
   product support
-version: 1.2.1
+version: 2.0.0
 homepage: https://github.com/cedtegApps/flutter_support_chat
 repository: https://github.com/cedtegApps/flutter_support_chat
 issue_tracker: https://github.com/cedtegApps/flutter_support_chat/issues
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
-  adaptive_dialog: ">=1.6.0 <2.0.0"
-  cloud_firestore: ">=2.2.0 <5.0.0"
-  device_info_plus: ">=7.0.0 <9.0.0"
+  adaptive_dialog: ">=2.0.0 <3.0.0"
+  cloud_firestore: ">=4.0.0 <5.0.0"
+  device_info_plus: ">=9.0.0 <10.0.0"
   firebase_core: ">=1.2.0 <3.0.0"
   flutter:
     sdk: flutter
